compiled

+++ server/migrations/2022-07-18-132458_txs/up.sql
@@ -7,7 +7,7 @@ CREATE TABLE txs (
     user_id BIGINT NOT NULL,
     eth_account BYTEA NOT NULL,
     usd_amount BIGINT NOT NULL,
-    crypto_amount BIGINT NOT NULL,
+    crypto_amount NUMERIC(39) NOT NULL,
     bid_date BIGINT NOT NULL,
     status txs_status_type NOT NULL DEFAULT 'before_ordered',
     tx_id BYTEA NULL,
